requirements.txt
Django==1.10.4
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.5.3
pkg-resources==0.0.0
requests==2.12.3
